S3: Changing our aiming mechanic
We are initially tracing from the crosshair's world location to the
point in space we are aiming at. If there is a trace hit then we set our beam end point to that location

We also need to perform a second line trace in order to check if there is anything between the gun barrel and that hit point and if  there is we need to update the beam end point to the new location representing the object intercepted by the
trace from the gun barrel to that object in-between.