S4: Initial Zoom functionality
Using simple functions to change the field of view of the UCameraComponent (we are calling FollowCamera). We are snap zooming right now (bound to the left ctrl button, when pressed = zoom in and when released = zoom out).  We need smooth interpolation so we can smooth out the zoom in/zoom out actions