S4: Crosshair spread based on velocity now implemented
We have a getter which is blueprint callable so we can get our new
crosshair spread multiplier value - Now expanded our HUD blueprint to get our character's velocity and then add/subtract x/y values for 4 separate images representing the arms of our crosshairs.  The crosshair spread multiplier is between 0.5 and 1.5 so this is multiplied by a blueprint variable called CrosshairSpreadMax which makes the spread more visible on screen.