S4: Further crosshair spread functionality implemented and aiming issue fixed
We have now implemented other factors to affect the crosshair spread (i.e. the inair factor based on if the character is in the air along with a small, quick shrink when aiming along with a quick expansion and contraction of the crosshairs when shooting) - When firing the weapon, i.e. when calling FireWeapon(),  StartCrosshairBulletFire() is then called which uses sets a boolean to interpolate crosshair movement and sets the boolean to false on a timer ..... Also, not using an y offset for the crosshair anymore as when aiming the position changes -  taking the offset out of GetBeamEndLocation and also out of the HUD BP.